Update patch set 1
Patch Set 1:

There would be a different way to do this.
Add BOARD_USES_FLUENCE_INCALL: = true and "patch" the snd_soc_msm_2x_Fusion3 file to "map" the mic devices to the ones that will be selected by the code then. IMHO this solution is not good cause it actually is "obfuscating" the problem

Patch-set: 1