Update patch set 1
Patch Set 1:

Ok, so I did a LOT of digging around and tried about 30 builds..

This is the first commit for the separated platform hals, from the kk_3.5 branch, and we need it for the voip mic to work on jflte (and probably other devices).

This commit alone does not fix the voip mic issue, as it'll need other, but I didn't want to waste my time if there's no chance this(or a similar replacement) will be merged in.

Patch-set: 1
Reviewer: Gerrit User 1000124 <1000124@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=0