Update patch set 3
Patch Set 3:

The only reason why moving the code makes a diff is that the
if (mCallMode == AUDIO_MODE_IN_COMMUNICATION) is cought somewhere before which you must prevent if you want the #ifdef SEPERATED_VOIP to be used on the original place. Just add a log line on every place where a different if path could be used then you see immediately which code path is used.

Patch-set: 3